# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Avenick Mobile API",
    "version": "1.0.0",
    "description": "The versioned surface the Avenick Flutter app consumes. Generated from the zod schemas in @avenick/contracts — never edited by hand. Every response is `{ data, meta? }`; every failure is `{ error: { code, message, requestId, fieldErrors? } }`."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Same-origin. /v1 paths are relative to this."
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from POST /v1/auth/token, /refresh or /otp/verify."
      }
    },
    "schemas": {
      "Image": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2048,
            "format": "uri"
          },
          "width": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 20000
          },
          "height": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 20000
          },
          "blurhash": {
            "type": "string",
            "minLength": 6,
            "maxLength": 128
          },
          "alt": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 300
          }
        },
        "required": [
          "url",
          "width",
          "height"
        ],
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "unauthenticated",
              "forbidden",
              "not_found",
              "validation_failed",
              "rate_limited",
              "conflict",
              "payment_required",
              "upstream_unavailable",
              "internal"
            ]
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "minItems": 1
            }
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "additionalProperties": false
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "ShippingAddress": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "line1": {
            "type": "string",
            "minLength": 3,
            "maxLength": 240
          },
          "line2": {
            "type": "string",
            "maxLength": 240
          },
          "city": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "country": {
            "type": "string",
            "enum": [
              "AE",
              "SA",
              "QA",
              "KW",
              "OM",
              "BH"
            ]
          },
          "postalCode": {
            "type": "string",
            "maxLength": 20
          }
        },
        "required": [
          "label",
          "line1",
          "city",
          "country"
        ],
        "additionalProperties": false
      },
...
      "PersistedOrderTotals": {
        "type": "object",
        "properties": {
          "subtotal": {
            "type": "number",
            "minimum": 0,
            "maximum": 9999999999.99
          },
          "discountAmount": {
            "type": "number",
            "minimum": 0,
            "maximum": 9999999999.99
          },
          "shippingAmount": {
            "type": "number",
            "minimum": 0,
            "maximum": 9999999999.99
          },
          "vatAmount": {
            "type": "number",
            "minimum": 0,
            "maximum": 9999999999.99
          },
          "goodsVatAmount": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "maximum": 9999999999.99
          },
          "shippingVatAmount": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "maximum": 9999999999.99
          },
          "total": {
            "type": "number",
            "minimum": 0,
            "maximum": 9999999999.99
          }
        },
        "required": [
          "subtotal",
          "discountAmount",
          "shippingAmount",
          "vatAmount",
          "goodsVatAmount",
          "shippingVatAmount",
          "total"
        ],
        "additionalProperties": false
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "productId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "variantId": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "maxLength": 128
          },
          "sellerId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "slug": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "maxLength": 160,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "sku": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "nameEn": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "nameAr": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "image": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Image"
              },
              {
                "type": [
                  "object",
                  "null"
                ]
              }
            ]
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "unitPrice": {
            "type": "number",
            "minimum": 0,
            "maximum": 9999999999.99
          },
          "vatRatePercent": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "vatAmount": {
            "type": "number",
            "minimum": 0,
            "maximum": 9999999999.99
          },
          "total": {
            "type": "number",
            "minimum": 0,
            "maximum": 9999999999.99
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING_PAYMENT",
              "PAYMENT_CONFIRMED",
              "CONFIRMED",
              "PROCESSING",
              "SHIPPED",
              "OUT_FOR_DELIVERY",
              "DELIVERED",
              "CANCELLED",
              "REFUNDED",
              "RETURN_REQUESTED",
              "RETURNED"
            ]
          }
        },
        "required": [
          "id",
          "productId",
          "variantId",
          "sellerId",
          "slug",
          "sku",
          "nameEn",
          "nameAr",
          "image",
          "quantity",
          "unitPrice",
          "vatRatePercent",
          "vatAmount",
          "total",
          "status"
        ],
        "additionalProperties": false
      },
      "OrderStatusEvent": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING_PAYMENT",
              "PAYMENT_CONFIRMED",
              "CONFIRMED",
              "PROCESSING",
              "SHIPPED",
              "OUT_FOR_DELIVERY",
              "DELIVERED",
              "CANCELLED",
              "REFUNDED",
              "RETURN_REQUESTED",
              "RETURNED"
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "status",
          "message",
          "occurredAt"
        ],
        "additionalProperties": false
      },
      "Shipment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PICKED_UP",
              "IN_TRANSIT",
              "OUT_FOR_DELIVERY",
              "DELIVERED",
              "FAILED",
              "RETURNED"
            ]
          },
          "carrier": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 120
          },
          "trackingNumber": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 120
          },
          "trackingUrl": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2048,
            "format": "uri"
          },
          "estimatedDelivery": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "status",
          "carrier",
          "trackingNumber",
          "trackingUrl",
          "estimatedDelivery"
        ],
        "additionalProperties": false
      },
...
      "OrderDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "orderNumber": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING_PAYMENT",
              "PAYMENT_CONFIRMED",
              "CONFIRMED",
              "PROCESSING",
              "SHIPPED",
              "OUT_FOR_DELIVERY",
              "DELIVERED",
              "CANCELLED",
              "REFUNDED",
              "RETURN_REQUESTED",
              "RETURNED"
            ]
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "UNPAID",
              "PAID",
              "PARTIALLY_PAID",
              "REFUNDED",
              "FAILED"
            ]
          },
          "paymentMethod": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "MADA",
              "APPLE_PAY",
              "CREDIT_CARD",
              "BANK_TRANSFER",
              "STC_PAY",
              "MOCK"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "B2C",
              "B2B"
            ]
          },
          "currency": {
            "type": "string",
            "enum": [
              "AED",
              "SAR",
              "QAR",
              "KWD",
              "OMR",
              "BHD",
              "USD"
            ]
          },
          "totals": {
            "$ref": "#/components/schemas/PersistedOrderTotals"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "minItems": 1,
            "maxItems": 500
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/ShippingAddress"
          },
          "shipments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Shipment"
            },
            "maxItems": 50
          },
          "statusHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderStatusEvent"
            },
            "maxItems": 200
          },
          "notes": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "vatInvoiceUrl": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2048,
            "format": "uri"
          },
          "placedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "status",
          "paymentStatus",
          "paymentMethod",
          "type",
          "currency",
          "totals",
          "items",
          "shippingAddress",
          "shipments",
          "statusHistory",
          "notes",
          "vatInvoiceUrl",
          "placedAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
...
    "/v1/orders/{id}": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "One order, in full",
        "description": "totals.goodsVatAmount and totals.shippingVatAmount are nullable here, unlike on a quote: the Order table stores a single collapsed vatAmount and the two components are not persisted.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The full order.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/OrderDetail"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such resource, or none visible to this caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Throttled. Honour the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected fault. Report the requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}